requirements.txt
flask>=3.1.2
flask-cors>=6.0.1
gunicorn>=23.0.0
python-dotenv>=1.2.1
requests>=2.32.5
opencv-python>=4.8.0
